+++ packages/cli/uni-builder/src/types.ts
@@ -12,6 +12,8 @@ import type {
   RsbuildEntry,
   PromiseOrNot,
   RsbuildPluginAPI,
+  ArrayOrNot,
+  HtmlTagDescriptor,
 } from '@rsbuild/shared';
 import type { RsbuildConfig } from '@rsbuild/core';
 import type { PluginAssetsRetryOptions } from '@rsbuild/plugin-assets-retry';
@@ -31,6 +33,7 @@ import type {
   UniBuilderStartServerResult,
 } from './shared/devServer';
 import type { PluginSourceBuildOptions } from '@rsbuild/plugin-source-build';
+import type TerserPlugin from 'terser-webpack-plugin';
 
 export type CreateBuilderCommonOptions = {
   entry?: RsbuildEntry;
@@ -68,25 +71,32 @@ export type DisableSourceMapOption =
       css?: boolean;
     };
 
+export type ToolsDevServerConfig = ChainedConfig<{
+  before?: RequestHandler[];
+  after?: RequestHandler[];
+  client?: DevConfig['client'];
+  compress?: ServerConfig['compress'];
+  devMiddleware?: {
+    writeToDisk?: DevConfig['writeToDisk'];
+  };
+  liveReload?: boolean;
+  headers?: ServerConfig['headers'];
+  historyApiFallback?: ServerConfig['historyApiFallback'];
+  hot?: boolean;
+  https?: DevServerHttpsOptions;
+  setupMiddlewares?: DevConfig['setupMiddlewares'];
+  proxy?: ServerConfig['proxy'];
+}>;
+
+export type TerserPluginOptions = TerserPlugin.BasePluginOptions &
+  TerserPlugin.DefinedDefaultMinimizerAndOptions<TerserPlugin.TerserOptions>;
+
+export type ToolsTerserConfig = ChainedConfig<TerserPluginOptions>;
+
 export type UniBuilderExtraConfig = {
   tools?: {
     styledComponents?: false | PluginStyledComponentsOptions;
-    devServer?: ChainedConfig<{
-      before?: RequestHandler[];
-      after?: RequestHandler[];
-      client?: DevConfig['client'];
-      compress?: ServerConfig['compress'];
-      devMiddleware?: {
-        writeToDisk?: DevConfig['writeToDisk'];
-      };
-      liveReload?: boolean;
-      headers?: ServerConfig['headers'];
-      historyApiFallback?: ServerConfig['historyApiFallback'];
-      hot?: boolean;
-      https?: DevServerHttpsOptions;
-      setupMiddlewares?: DevConfig['setupMiddlewares'];
-      proxy?: ServerConfig['proxy'];
-    }>;
+    devServer?: ToolsDevServerConfig;
     /**
      * Configure the [Pug](https://pugjs.org/) template engine.
      */
@@ -113,6 +123,11 @@ export type UniBuilderExtraConfig = {
      * Tips: this configuration is not yet supported in rspack
      */
     tsLoader?: PluginTsLoaderOptions;
+    /**
+     * Modify the options of [terser-webpack-plugin](https://github.com/webpack-contrib/terser-webpack-plugin).
+     * @requires webpack
+     */
+    terser?: ToolsTerserConfig;
   };
   dev?: {
     /**
@@ -151,6 +166,11 @@ export type UniBuilderExtraConfig = {
     resolveExtensionPrefix?: string | Partial<Record<RsbuildTarget, string>>;
   };
   output?: {
+    /**
+     * Whether to disable code minification in production build.
+     */
+    // TODO: support output.minify configuration
+    disableMinimize?: boolean;
     /**
      * @deprecated use `output.filenameHash` instead
      */
@@ -223,6 +243,10 @@ export type UniBuilderExtraConfig = {
      * When this option is enabled, the generated HTML file path will change from `[name]/index.html` to `[name].html`.
      */
     disableHtmlFolder?: boolean;
+    /**
+     *  @deprecated use `html.tags` instead
+     */
+    tagsByEntries?: Record<string, ArrayOrNot<HtmlTagDescriptor>>;
     /**
      * @deprecated use `html.meta` instead
      */