+++ packages/cli/uni-builder/src/webpack/index.ts
@@ -57,6 +57,11 @@ export async function parseConfig(
     rsbuildPlugins.push(pluginManifest());
   }
 
+  if (!uniBuilderConfig.output?.disableMinimize) {
+    const { pluginMinimize } = await import('./plugins/minimize');
+    rsbuildPlugins.push(pluginMinimize(uniBuilderConfig.tools?.terser));
+  }
+
   if (uniBuilderConfig.security?.sri) {
     const { pluginSRI } = await import('./plugins/sri');
     rsbuildPlugins.push(pluginSRI(uniBuilderConfig.security?.sri));