+++ packages/server/server/src/types.ts
@@ -5,11 +5,7 @@ import type {
   NextFunction,
 } from '@modern-js/types';
 import type { ModernServerOptions } from '@modern-js/prod-server';
-import type {
-  RsbuildInstance,
-  DevServerAPIs,
-  DevMiddlewaresConfig,
-} from '@rsbuild/shared';
+import type { RsbuildInstance, RsbuildDevServer } from '@rsbuild/shared';
 
 export type { DevServerOptions, DevServerHttpsOptions };
 
@@ -61,12 +57,15 @@ export type ExtraOptions = {
 export type ModernDevServerOptions = ModernServerOptions & ExtraOptions;
 
 export type ExtraOptionsNew = {
-  dev: Partial<DevServerOptions>;
+  dev: Pick<DevServerOptions, 'watch' | 'https'> & {
+    writeToDisk?: boolean | ((filename: string) => boolean);
+  };
   useSSRWorker?: boolean;
   rsbuild: RsbuildInstance;
-  getMiddlewares: (
-    overrides?: DevMiddlewaresConfig,
-  ) => ReturnType<DevServerAPIs['getMiddlewares']>;
+  getMiddlewares: () => Pick<
+    RsbuildDevServer,
+    'middlewares' | 'onHTTPUpgrade' | 'close'
+  >;
 };
 
 export type ModernDevServerOptionsNew = ModernServerOptions & ExtraOptionsNew;