+++ tests/e2e/builder/cases/html-tags/index.test.ts
@@ -33,3 +33,59 @@ test('should inject tags', async () => {
   ).toBeTruthy();
   expect(isHtmlMatch(indexHtml, /content="origin"/)).toBeTruthy();
 });
+
+test('should inject tags by tagsByEntries', async () => {
+  const builder = await build({
+    cwd: __dirname,
+    entry: { index: path.resolve(__dirname, './src/index.ts') },
+    builderConfig: {
+      html: {
+        tags: [
+          tags => {
+            // should get foo tag info
+            if (tags.find(tag => tag.attrs?.src === 'foo.js')) {
+              tags.push({
+                tag: 'script',
+                attrs: { src: 'https://www.cdn.com/foo.js' },
+              });
+            }
+          },
+          { tag: 'script', attrs: { src: 'foo.js' } },
+        ],
+        tagsByEntries: {
+          index: [
+            tags => {
+              // should get bar tag info
+              if (tags.find(tag => tag.attrs?.src === 'bar.js')) {
+                tags.push({
+                  tag: 'meta',
+                  attrs: { name: 'referrer', content: 'origin' },
+                });
+              }
+            },
+
+            tags => [
+              ...tags,
+              { tag: 'script', attrs: { src: 'baz.js' }, append: false },
+            ],
+            { tag: 'script', attrs: { src: 'bar.js' }, append: false },
+          ],
+          index2: [{ tag: 'script', attrs: { src: '404.js' }, append: false }],
+        },
+      },
+    },
+  });
+
+  const files = await builder.unwrapOutputJSON();
+
+  const indexHtml =
+    files[path.resolve(__dirname, './dist/html/index/index.html')];
+
+  expect(isHtmlMatch(indexHtml, /bar\.js/)).toBeTruthy();
+  expect(isHtmlMatch(indexHtml, /foo\.js/)).toBeTruthy();
+  expect(
+    isHtmlMatch(indexHtml, /src="https:\/\/www\.cdn\.com\/foo\.js/),
+  ).toBeTruthy();
+  expect(isHtmlMatch(indexHtml, /content="origin"/)).toBeTruthy();
+  expect(isHtmlMatch(indexHtml, /404\.js/)).toBeFalsy();
+});