+++ packages/document/main-doc/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @modern-js/main-doc
 
+## 2.48.2
+
+### Patch Changes
+
+- @modern-js/builder-doc@2.48.2
+- @modern-js/sandpack-react@2.48.2
+
 ## 2.48.1
 
 ### Patch Changes