+++ packages/generator/generators/router-v5-generator/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js/router-v5-generator
 
+## 3.3.13
+
+### Patch Changes
+
+- @modern-js/utils@2.48.2
+
 ## 3.3.12
 
 ### Patch Changes