+++ packages/server/utils/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @modern-js/server-utils
 
+## 2.48.2
+
+### Patch Changes
+
+- @modern-js/babel-plugin-module-resolver@2.48.2
+- @modern-js/babel-compiler@2.48.2
+- @modern-js/utils@2.48.2
+
 ## 2.48.1
 
 ### Patch Changes